Reorganize bees module
