+++ src/java/growthcraft/bees/block/BlockBeeBox.java
@@ -1,14 +1,19 @@
-package growthcraft.bees;
-
-import growthcraft.api.bees.BeesRegistry;
-import growthcraft.core.GrowthCraftCore;
-import growthcraft.core.Utils;
+package growthcraft.bees.block;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
 
+import growthcraft.api.bees.BeesRegistry;
+import growthcraft.bees.entity.TileEntityBeeBox;
+import growthcraft.bees.GrowthCraftBees;
+import growthcraft.bees.renderer.RenderBeeBox;
+import growthcraft.core.GrowthCraftCore;
+import growthcraft.core.Utils;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockContainer;
 import net.minecraft.block.material.Material;
@@ -28,8 +33,6 @@ import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.ForgeDirection;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class BlockBeeBox extends BlockContainer
 {
