+++ src/java/growthcraft/bees/block/BlockBeeHive.java
@@ -1,10 +1,14 @@
-package growthcraft.bees;
-
-import growthcraft.core.GrowthCraftCore;
+package growthcraft.bees.block;
 
 import java.util.List;
 import java.util.Random;
 
+import growthcraft.core.GrowthCraftCore;
+import growthcraft.bees.GrowthCraftBees;
+import growthcraft.bees.renderer.RenderBeeHive;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IIconRegister;
@@ -19,8 +23,6 @@ import net.minecraft.util.IIcon;
 import net.minecraft.util.MathHelper;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class BlockBeeHive extends Block
 {
