+++ src/java/growthcraft/bees/block/ContainerBeeBox.java
@@ -1,4 +1,9 @@
-package growthcraft.bees;
+package growthcraft.bees.block;
+
+import growthcraft.bees.entity.TileEntityBeeBox;
+import growthcraft.bees.GrowthCraftBees;
+import growthcraft.bees.inventory.SlotBee;
+import growthcraft.bees.inventory.SlotHoneyComb;
 
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.InventoryPlayer;
