+++ src/java/growthcraft/bees/gui/GuiHandlerBees.java
@@ -1,9 +1,13 @@
-package growthcraft.bees;
+package growthcraft.bees.gui;
 
+import growthcraft.bees.block.ContainerBeeBox;
+import growthcraft.bees.entity.TileEntityBeeBox;
+import growthcraft.bees.gui.GuiBeeBox;
+
+import cpw.mods.fml.common.network.IGuiHandler;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
-import cpw.mods.fml.common.network.IGuiHandler;
 
 public class GuiHandlerBees implements IGuiHandler
 {
