+++ src/java/growthcraft/bees/inventory/SlotHoneyComb.java
@@ -1,4 +1,7 @@
-package growthcraft.bees;
+package growthcraft.bees.inventory;
+
+import growthcraft.bees.GrowthCraftBees;
+import growthcraft.bees.block.ContainerBeeBox;
 
 import net.minecraft.inventory.IInventory;
 import net.minecraft.inventory.Slot;
