+++ src/java/growthcraft/bees/item/ItemHoneyComb.java
@@ -1,17 +1,17 @@
-package growthcraft.bees;
-
-import growthcraft.core.GrowthCraftCore;
+package growthcraft.bees.item;
 
 import java.util.List;
 
+import growthcraft.core.GrowthCraftCore;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.IIcon;
 import net.minecraft.util.MathHelper;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class ItemHoneyComb extends Item
 {
