+++ src/java/growthcraft/bees/renderer/RenderBeeHive.java
@@ -1,13 +1,15 @@
-package growthcraft.bees;
+package growthcraft.bees.renderer;
 
+import growthcraft.bees.block.BlockBeeHive;
 import growthcraft.core.Utils;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderBeeHive implements ISimpleBlockRenderingHandler
 {
