+++ src/java/growthcraft/bees/world/WorldGeneratorBees.java
@@ -1,15 +1,17 @@
-package growthcraft.bees;
-
-import growthcraft.core.Utils;
+package growthcraft.bees.world;
 
 import java.util.Random;
 
+import growthcraft.bees.GrowthCraftBees;
+import growthcraft.bees.world.WorldGenBeeHive;
+import growthcraft.core.Utils;
+
+import cpw.mods.fml.common.IWorldGenerator;
 import net.minecraft.world.World;
 import net.minecraft.world.biome.BiomeGenBase;
 import net.minecraft.world.chunk.IChunkProvider;
 import net.minecraftforge.common.BiomeDictionary;
 import net.minecraftforge.common.BiomeDictionary.Type;
-import cpw.mods.fml.common.IWorldGenerator;
 
 public class WorldGeneratorBees implements IWorldGenerator
 {
